Fixed imports, moved functions to better locations and fixed some codacy errors
