+++ vcstools/general_utils.py
@@ -9,6 +9,21 @@ from astropy.utils import iers
 
 logger = logging.getLogger(__name__)
 
+def sfreq(freqs):
+
+    if len(freqs) != 24:
+        print("There are not 24 coarse chans defined for this obs. Got: %s" % freqs)
+        return
+
+    #freqs.sort()   # It should already be sorted, but just in case...[SET] Commenting this out because sort() is ironically putting 2-digit channels out of order
+    lowchans = [f for f in freqs if f <= 128]
+    highchans = [f for f in freqs if f > 128]
+
+    highchans.reverse()
+    freqs = lowchans + highchans
+
+    return freqs
+
 
 def is_number(s):
     try:
